+++ packages/flutter_fuelet_smart_contract_wallet/lib/wallet/wrapped_smart_contract_wallet.dart
@@ -1,0 +1,17 @@
+abstract class WrappedSmartContractWallet {
+  String get predicateBech32Address;
+
+  String get r1PublicKey;
+
+  Future<void> deployContract();
+
+  Future<(String, String)> createTransferTxRequest({
+    required String destinationB256Address,
+    required int fractionalAmount,
+    required String assetId,
+  });
+
+  Future<String> sendTransaction(
+      {required String transactionRequestHexOrJson,
+      required String r1Signature});
+}