+++ CHANGELOG.md
@@ -15,6 +15,133 @@ Packages with breaking changes:
 
 Packages with other changes:
 
+ - [`fuelet_smart_contract_wallet` - `v1.2.3`](#fuelet_smart_contract_wallet---v123)
+ - [`flutter_fuelet_smart_contract_wallet` - `v0.0.8`](#flutter_fuelet_smart_contract_wallet---v008)
+
+---
+
+#### `fuelet_smart_contract_wallet` - `v1.2.3`
+
+ - Bump "fuelet_smart_contract_wallet" to `1.2.3`.
+
+#### `flutter_fuelet_smart_contract_wallet` - `v0.0.8`
+
+ - Bump "flutter_fuelet_smart_contract_wallet" to `0.0.8`.
+
+
+## 2024-02-06
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`fuelet_smart_contract_wallet` - `v1.2.2`](#fuelet_smart_contract_wallet---v122)
+ - [`flutter_fuelet_smart_contract_wallet` - `v0.0.7`](#flutter_fuelet_smart_contract_wallet---v007)
+
+---
+
+#### `fuelet_smart_contract_wallet` - `v1.2.2`
+
+ - Bump "fuelet_smart_contract_wallet" to `1.2.2`.
+
+#### `flutter_fuelet_smart_contract_wallet` - `v0.0.7`
+
+ - Bump "flutter_fuelet_smart_contract_wallet" to `0.0.7`.
+
+
+## 2024-02-06
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`fuelet_smart_contract_wallet` - `v1.2.1`](#fuelet_smart_contract_wallet---v121)
+ - [`flutter_fuelet_smart_contract_wallet` - `v0.0.6`](#flutter_fuelet_smart_contract_wallet---v006)
+
+---
+
+#### `fuelet_smart_contract_wallet` - `v1.2.1`
+
+ - Fix build scripts
+
+#### `flutter_fuelet_smart_contract_wallet` - `v0.0.6`
+
+ - Fix build scripts
+
+
+## 2024-02-06
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`flutter_fuelet_smart_contract_wallet` - `v0.0.5`](#flutter_fuelet_smart_contract_wallet---v005)
+
+---
+
+#### `flutter_fuelet_smart_contract_wallet` - `v0.0.5`
+
+ - **FIX**: podspec. ([1707b67c](https://github.com/Fuelet/SmartContractWalletSDK/commit/1707b67cb6b89609a2f33f14fb01f5fe0fb5b8cc))
+ - **FIX**: mobile wallet bindings. ([2fd5b041](https://github.com/Fuelet/SmartContractWalletSDK/commit/2fd5b041bc3fd97b914d27e8f36d5964ee018432))
+ - **FEAT**: basic example. ([a8d886b4](https://github.com/Fuelet/SmartContractWalletSDK/commit/a8d886b40e38f3a4425f9c4b28244b98cd560da5))
+
+
+## 2024-02-06
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`flutter_fuelet_smart_contract_wallet` - `v0.0.4`](#flutter_fuelet_smart_contract_wallet---v004)
+ - [`fuelet_smart_contract_wallet` - `v1.2.0`](#fuelet_smart_contract_wallet---v120)
+
+---
+
+#### `flutter_fuelet_smart_contract_wallet` - `v0.0.4`
+
+ - **FEAT**: flutter smart contract wallet implementation. ([cd758597](https://github.com/Fuelet/SmartContractWalletSDK/commit/cd75859773691a3646d8843ffc2dd153c8798f81))
+
+#### `fuelet_smart_contract_wallet` - `v1.2.0`
+
+ - **FEAT**: flutter smart contract wallet implementation. ([cd758597](https://github.com/Fuelet/SmartContractWalletSDK/commit/cd75859773691a3646d8843ffc2dd153c8798f81))
+
+
+## 2024-02-06
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
  - [`flutter_fuelet_smart_contract_wallet` - `v0.0.3`](#flutter_fuelet_smart_contract_wallet---v003)
  - [`fuelet_smart_contract_wallet` - `v1.1.0`](#fuelet_smart_contract_wallet---v110)