+++ packages/fuelet_smart_contract_wallet/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.2.3
+
+ - Bump "fuelet_smart_contract_wallet" to `1.2.3`.
+
+## 1.2.2
+
+ - Bump "fuelet_smart_contract_wallet" to `1.2.2`.
+
+## 1.2.1
+
+ - Fix build scripts
+
+## 1.2.0
+
+ - **FEAT**: flutter smart contract wallet implementation. ([cd758597](https://github.com/Fuelet/SmartContractWalletSDK/commit/cd75859773691a3646d8843ffc2dd153c8798f81))
+
 ## 1.1.0
 
  - **FEAT**: contracts interaction ([#1](https://github.com/Fuelet/SmartContractWalletSDK/issues/1)). ([00a4592f](https://github.com/Fuelet/SmartContractWalletSDK/commit/00a4592f0da26750d531466bd11af06331b568f3))